pyproject.toml
[project]
name = "gemini-quick-start"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi[standard]>=0.135.1",
    "google-adk>=1.10.0",
    "google-genai>=1.66.0",
    "litellm>=1.82.4",
    "logging>=0.4.9.6",
    "numpy>=2.4.2",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "websockets>=16.0",
]
